@@ -1,112 +1,4 @@
-<template>
-  <div>
-    <div v-if="token.length < 1">
-      <div class="panel-section panel-section-header">
-        <p>Please provide your Pinboard API token in the options.</p>
-      </div>
-      <div class="panel-section panel-section-footer">
-        <div class="panel-section-footer-button default" @click="openOptions()">
-          Open Epic Pinboard options
-        </div>
-      </div>
-    </div>
-    <div v-else>
-      <div v-show="loading">
-        <div class="ball-pulse">
-          <div></div>
-          <div></div>
-          <div></div>
-        </div>
-      </div>
-      <div class="panel-section panel-section-formElements" v-show="!loading">
-        <form id="addBoomarkForm">
-          <div class="form-item browser-style">
-            <label for="url">URL</label>
-            <input
-              type="text"
-              name="url"
-              id="url"
-              maxlength="2000"
-              v-model="url"
-            />
-          </div>
-          <div class="form-item browser-style">
-            <label for="title">Title</label>
-            <input
-              type="text"
-              name="title"
-              id="title"
-              maxlength="255"
-              v-model="title"
-            />
-          </div>
-          <div class="form-item browser-style">
-            <label for="tags">Tags</label>
-            <vue-tags-input
-              ref="tagsInput"
-              v-model="tag"
-              :tags="tags"
-              :autocomplete-items="filteredTags"
-              @tags-changed="(newTags) => (tags = newTags)"
-            />
-          </div>
-          <div class="form-item browser-style">
-            <label for="suggested">Suggested Tags</label>
-            <ul class="suggested-tags">
-              <li
-                class="suggested-tag"
-                v-for="tag in suggestedTags"
-                :key="tag"
-                @click="addSuggestedTag(tag)"
-              >
-                {{ tag }}
-              </li>
-            </ul>
-          </div>
-          <div class="form-item browser-style">
-            <label for="notes">Notes</label>
-            <textarea
-              class="browser-style"
-              name="notes"
-              id="notes"
-              maxlength="65536"
-              v-model="notes"
-            ></textarea>
-          </div>
-          <div class="form-item browser-style">
-            <input
-              type="checkbox"
-              id="private"
-              name="private"
-              v-model="privateBookmark"
-            />
-            <label for="private" class="inline">Private bookmark</label>
-          </div>
-          <div class="form-item browser-style">
-            <input
-              type="checkbox"
-              id="readlater"
-              name="readlater"
-              v-model="readLater"
-            />
-            <label for="readlater" class="inline">Read later</label>
-          </div>
-        </form>
-      </div>
-      <div class="panel-section panel-section-footer" v-show="!loading">
-        <div
-          class="panel-section-footer-button default"
-          id="addBookmarkButton"
-          @click="addBookmark"
-        >
-          {{ buttonCaption }}
-        </div>
-      </div>
-    </div>
-  </div>
-</template>
-
-<script>
+<script setup>
 import { VueTagsInput, createTags } from "@johmun/vue-tags-input";
 
 function getActiveTab() {
@@ -121,9 +13,6 @@ const ctaAddBookmark = "Add bookmark";
 const ctaUpdateBookmark = "Update bookmark";
 
 export default {
-  components: {
-    VueTagsInput,
-  },
   data() {
     return {
       token: "",
@@ -263,6 +152,114 @@ export default {
   },
 };
 </script>
+
+<template>
+  <div>
+    <div v-if="token.length < 1">
+      <div class="panel-section panel-section-header">
+        <p>Please provide your Pinboard API token in the options.</p>
+      </div>
+      <div class="panel-section panel-section-footer">
+        <div class="panel-section-footer-button default" @click="openOptions()">
+          Open Epic Pinboard options
+        </div>
+      </div>
+    </div>
+    <div v-else>
+      <div v-show="loading">
+        <div class="ball-pulse">
+          <div></div>
+          <div></div>
+          <div></div>
+        </div>
+      </div>
+      <div class="panel-section panel-section-formElements" v-show="!loading">
+        <form id="addBoomarkForm">
+          <div class="form-item browser-style">
+            <label for="url">URL</label>
+            <input
+              type="text"
+              name="url"
+              id="url"
+              maxlength="2000"
+              v-model="url"
+            />
+          </div>
+          <div class="form-item browser-style">
+            <label for="title">Title</label>
+            <input
+              type="text"
+              name="title"
+              id="title"
+              maxlength="255"
+              v-model="title"
+            />
+          </div>
+          <div class="form-item browser-style">
+            <label for="tags">Tags</label>
+            <vue-tags-input
+              ref="tagsInput"
+              v-model="tag"
+              :tags="tags"
+              :autocomplete-items="filteredTags"
+              @tags-changed="(newTags) => (tags = newTags)"
+            />
+          </div>
+          <div class="form-item browser-style">
+            <label for="suggested">Suggested Tags</label>
+            <ul class="suggested-tags">
+              <li
+                class="suggested-tag"
+                v-for="tag in suggestedTags"
+                :key="tag"
+                @click="addSuggestedTag(tag)"
+              >
+                {{ tag }}
+              </li>
+            </ul>
+          </div>
+          <div class="form-item browser-style">
+            <label for="notes">Notes</label>
+            <textarea
+              class="browser-style"
+              name="notes"
+              id="notes"
+              maxlength="65536"
+              v-model="notes"
+            ></textarea>
+          </div>
+          <div class="form-item browser-style">
+            <input
+              type="checkbox"
+              id="private"
+              name="private"
+              v-model="privateBookmark"
+            />
+            <label for="private" class="inline">Private bookmark</label>
+          </div>
+          <div class="form-item browser-style">
+            <input
+              type="checkbox"
+              id="readlater"
+              name="readlater"
+              v-model="readLater"
+            />
+            <label for="readlater" class="inline">Read later</label>
+          </div>
+        </form>
+      </div>
+      <div class="panel-section panel-section-footer" v-show="!loading">
+        <div
+          class="panel-section-footer-button default"
+          id="addBookmarkButton"
+          @click="addBookmark"
+        >
+          {{ buttonCaption }}
+        </div>
+      </div>
+    </div>
+  </div>
+</template>
 
 <style>
 .suggested-tags {
